requirements.txt
python-telegram-bot==20.7
requests==2.31.0
Pillow==10.0.1
SpeechRecognition==3.10.0
pydub==0.25.1
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
nest-asyncio==1.5.8
urllib3>=1.26.0
